[REFACTOR] Divides OS Places API client into interface and implementation
