+++ src/main/kotlin/uk/gov/justice/digital/hmpps/jobsboard/api/jobs/infrastructure/OSPlacesAPIWebClient.kt
@@ -1,10 +1,11 @@
 package uk.gov.justice.digital.hmpps.jobsboard.api.jobs.infrastructure
 
 import org.springframework.stereotype.Service
+import uk.gov.justice.digital.hmpps.jobsboard.api.jobs.domain.OSPlacesAPIClient
 
 @Service
-class OSPlacesAPIClient {
-  fun getAddressesFor(postcode: String): OSPlacesAPIDPA {
+class OSPlacesAPIWebClient : OSPlacesAPIClient {
+  override fun getAddressesFor(postcode: String): OSPlacesAPIDPA {
     return OSPlacesAPIDPA(
       postcode = "",
       xCoordinate = 0.00,
